refactored global data call to own method, set up userquerydata method and conditions
